+++ sdk/platform/Communication/CommunicationPlatformApplicationValidator.js
@@ -176,6 +176,14 @@ const CommunicationPlatformModel = require("./CommunicationPlatformModel");
  * @property {number} [pageNo] - Current page no
  * @property {number} [pageSize] - Current request items count
  * @property {string} [populate] - Populate Fields
+ * @property {string} [group] - An event group is a collection of email and SMS
+ *   templates. Filtering by event group lets you view or manage all related
+ *   communication templates together.
+ * @property {string} [subGroup] - Filter by event subgroup. Here, a subgroup is
+ *   a subset within a group, containing specific email and SMS templates for
+ *   more detailed organization.
+ * @property {string} [fulfillmentOptionTypes] - Filter by fulfillment option
+ *   type. Indicates the delivery choice selected, e.g., standard or express.
  */
 
 /**
@@ -587,6 +595,9 @@ class CommunicationPlatformApplicationValidator {
       pageNo: Joi.number(),
       pageSize: Joi.number(),
       populate: Joi.string().allow(""),
+      group: Joi.string().allow(""),
+      subGroup: Joi.string().allow(""),
+      fulfillmentOptionTypes: Joi.string().allow(""),
     }).required();
   }