+++ sdk/platform/Content/ContentPlatformApplicationValidator.js
@@ -360,10 +360,6 @@ const ContentPlatformModel = require("./ContentPlatformModel");
 /**
  * @typedef GetInjectableTagsParam
  * @property {boolean} [all] - Get all tags irrespective of the creator of tags
- * @property {number} [pageNo] - The page number to navigate through the given
- *   set of results. Default value is 1.
- * @property {number} [pageSize] - The number of items to retrieve in each page.
- *   Default value is 10.
  * @property {string} [search] - Keyword to filter and find tags by name.
  */
 
@@ -1097,8 +1093,6 @@ class ContentPlatformApplicationValidator {
   static getInjectableTags() {
     return Joi.object({
       all: Joi.boolean(),
-      pageNo: Joi.number(),
-      pageSize: Joi.number(),
       search: Joi.string().allow(""),
     }).required();
   }